rename the exception and move it out of BitmapDecodeException
